requirements.txt
Flask
pillow~=8.1.2
opencv-python~=4.5.1.48
requests~=2.25.1
libtiff~=0.4.2
keras~=2.4.3
pyproj~=2.6.1.post1
tensorflow~=2.3.0
matplotlib~=3.3.4
numpy~=1.19.2